Add support for URL percent decoding (RFC 1738)
Initial support and fairly shallow testing of it.
Spec says anything that is not US-ASCII should be encoded, but does not
specify the encoding that should be supplied ‒ which I assume is now
down to application use. For sake of simplicity this opts for UTF-8 but
I can forsee needing to expand this to allow the user to request the
input charset. It will always return UTF-8.